# modules/apps/data-engine/data-engine-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        DataRecord:
            description:
                https://www.schema.org/DataRecord
            properties:
                dataRecordCollectionId:
                    format: int64
                    type: integer
                dataRecordValues:
                    additionalProperties:
                        type: object
                    type: object
                id:
                    format: int64
                    type: integer
        DataRecordCollection:
            description:
                https://www.schema.org/DataRecordCollection
            properties:
                dataDefinitionId:
                    format: int64
                    type: integer
                dataRecordCollectionKey:
                    type: string
                description:
                    additionalProperties:
                        type: object
                    type: object
                id:
                    format: int64
                    type: integer
                name:
                    additionalProperties:
                        type: object
                    type: object
                siteId:
                    format: int64
                    type: integer
            type: object
        DataRecordCollectionPermission:
            properties:
                addDataRecord:
                    type: boolean
                addDataRecordCollection:
                    type: boolean
                definePermissions:
                    type: boolean
                delete:
                    type: boolean
                deleteDataRecord:
                    type: boolean
                exportDataRecord:
                    type: boolean
                roleNames:
                    items:
                        type: string
                    type:
                        array
                update:
                    type: boolean
                updateDataRecord:
                    type: boolean
                view:
                    type: boolean
                viewDataRecord:
                    type: boolean
            type: object
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Data Engine"
    version: v2.0
openapi: 3.0.1
paths:
...
    "/data-definitions/{dataDefinitionId}/data-record-collections":
        get:
            operationId: getDataDefinitionDataRecordCollectionsPage
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecordCollection"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecordCollection"
                                type: array
                    description:
                        ""
            tags: ["DataRecordCollection"]
        post:
            operationId: postDataDefinitionDataRecordCollection
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollection"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollection"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                    description:
                        ""
            tags: ["DataRecordCollection"]
    "/data-definitions/{dataDefinitionId}/data-records":
        get:
            operationId: getDataDefinitionDataRecordsPage
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                    description:
                        ""
            tags: ["DataRecord"]
        post:
            operationId: postDataDefinitionDataRecord
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
...
    "/data-record-collections/{dataRecordCollectionId}":
        delete:
            operationId: deleteDataRecordCollection
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description:
                        ""
            tags: ["DataRecordCollection"]
        get:
            operationId: getDataRecordCollection
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                    description:
                        ""
            tags: ["DataRecordCollection"]
        put:
            operationId: putDataRecordCollection
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollection"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollection"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                    description:
                        ""
            tags: ["DataRecordCollection"]
    "/data-record-collections/{dataRecordCollectionId}/data-record-collection-permissions":
        post:
            operationId: postDataRecordCollectionDataRecordCollectionPermission
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: operation
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollectionPermission"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollectionPermission"
                description:
                    ""
            responses:
                204:
                    content:
                        application/json: {}
                    description:
                        ""
            tags: ["DataRecordCollection"]
    "/data-record-collections/{dataRecordCollectionId}/data-records":
        get:
            operationId: getDataRecordCollectionDataRecordsPage
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                    description:
                        ""
            tags: ["DataRecord"]
        post:
            operationId: postDataRecordCollectionDataRecord
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
...
    "/data-records/{dataRecordId}":
        delete:
            operationId: deleteDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description:
                        ""
            tags: ["DataRecord"]
        get:
            operationId: getDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
        put:
            operationId: putDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
...
    "/sites/{siteId}/data-record-collection-permissions":
        post:
            operationId: postSiteDataRecordCollectionPermission
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: operation
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollectionPermission"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollectionPermission"
            responses:
                204:
                    content:
                        application/json: {}
                    description:
                        ""
            tags: ["DataRecordCollection"]
    "/sites/{siteId}/data-record-collections":
        get:
            operationId: getSiteDataRecordCollectionsPage
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecordCollection"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecordCollection"
                                type: array
                    description:
                        ""
            tags: ["DataRecordCollection"]
    "/sites/{siteId}/data-record-collections/{dataRecordCollectionKey}":
        get:
            operationId: getSiteDataRecordCollection
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: dataRecordCollectionKey
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                    description:
                        ""
            tags: ["DataRecordCollection"]